quick change

+++ app/src/main/java/data/FirebaseDataModels.kt
@@ -55,8 +55,8 @@ data class Group(val groupName:String,
 
 data class Message(val sendingId:String,
                    val msg:String,
-                   val timestamp:Int,
+                   val timestamp:Long,
                    val read:Boolean=false,
                    @get:Exclude override val uid: String?=null): BaseData() {
-    constructor() : this("", "", 0, false, null)
+    constructor() : this("", "", 0L, false, null)
 }
